build: validate resources+lenght on post consolidated
Ref. metriport/metriport-internal#870
