+++ packages/api/src/routes/medical/schemas/fhir.ts
@@ -1,4 +1,12 @@
+import { Request } from "express";
 import { z } from "zod";
+import BadRequestError from "../../../errors/bad-request";
+import {
+  ResourceTypeForConsolidation,
+  resourceTypeForConsolidation,
+} from "../../../external/fhir/patient/consolidated";
+import { filterTruthy } from "../../../shared/filter-map-utils";
+import { getFrom } from "../../util";
 
 const typeSchema = z.enum(["collection"]);
 
@@ -16,3 +24,23 @@ export const bundleSchema = z.object({
 
 export type BundleEntry = z.infer<typeof bundleEntrySchema>;
 export type Bundle = z.infer<typeof bundleSchema>;
+
+export const resourceSchema = z.enum(resourceTypeForConsolidation).array();
+
+export function getResourcesQueryParam(req: Request): ResourceTypeForConsolidation[] {
+  const resourcesRaw = getFrom("query").optional("resources", req);
+  let resourcesUnparsed: string[];
+  try {
+    resourcesUnparsed = resourcesRaw
+      ? resourcesRaw
+          .replaceAll('"', "")
+          .replaceAll("'", "")
+          .split(",")
+          .map(r => r.trim())
+          .flatMap(filterTruthy)
+      : [];
+  } catch (err) {
+    throw new BadRequestError(`Invalid resources: it must be a comma-separated list of resources`);
+  }
+  return resourceSchema.parse(resourcesUnparsed);
+}
